src/main/java/ru/temnenkov/jnoderest/dto/Message.java
package ru.temnenkov.jnoderest.dto;

import java.util.Objects;

public class Message {
    private String echoArea;
    private String subject;
    private String body;

    public String getEchoArea() {
        return echoArea;
    }

    public void setEchoArea(String echoArea) {
        this.echoArea = echoArea;
    }

    public String getSubject() {
        return subject;
    }

    public void setSubject(String subject) {
        this.subject = subject;
    }

    public String getBody() {
        return body;
    }

    public void setBody(String body) {
        this.body = body;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Message message = (Message) o;
        return Objects.equals(echoArea, message.echoArea) &&
                Objects.equals(subject, message.subject) &&
                Objects.equals(body, message.body);
    }

    @Override
    public int hashCode() {
        return Objects.hash(echoArea, subject, body);
    }

    @Override
    public String toString() {
        final StringBuilder sb = new StringBuilder("Message{");
        sb.append("echoArea='").append(echoArea).append('\'');
        sb.append(", subject='").append(subject).append('\'');
        sb.append(", body='").append(body).append('\'');
        sb.append('}');
        return sb.toString();
    }
}
